changed spectra to float, calculating log coefficients and log binomial probability, convert back using exponent - much faster & can use on any sample size, limited to no floating point accumulation error, for the case of multiple populations added a if statement that if p = 1 or 0, the contribution applies only to the monomorphic sites

+++ include/spectrum.h
@@ -12,7 +12,7 @@
 namespace SPECTRUM{
 
 struct sfs{
-	double * frequency_spectrum;
+	float * frequency_spectrum;
 	int * populations; //which populations are in SFS
 	int * sample_size; //number of samples taken for each population
 	int num_populations;
